+++ feature/onboarding/src/main/java/com/yapp/onboarding/OnboardingCompleteScreen1.kt
@@ -1,22 +1,16 @@
-package com.kms.onboarding
+package com.yapp.onboarding
 
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.imePadding
-import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.res.stringResource
@@ -24,59 +18,43 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.kms.onboarding.component.OnBoardingTopAppBar
 import com.yapp.designsystem.theme.OrbitTheme
-import com.yapp.ui.component.button.OrbitButton
 import com.yapp.ui.component.lottie.LottieAnimation
 import com.yapp.ui.utils.heightForScreenPercentage
 import feature.onboarding.R
 
 @Composable
-fun OnboardingCompleteRoute2(
+fun OnboardingCompleteRoute(
     viewModel: OnboardingViewModel,
 ) {
     val state by viewModel.container.stateFlow.collectAsStateWithLifecycle()
     BackHandler {
         viewModel.processAction(OnboardingContract.Action.PreviousStep) // ✅ ViewModel에서 처리
     }
-    OnboardingCompleteScreen2(
+    OnboardingCompleteScreen1(
         state = state,
-        onNextClick = {
-            viewModel.processAction(OnboardingContract.Action.CompleteOnboarding)
-            viewModel.processAction(OnboardingContract.Action.CreateAlarm)
-        },
+        onNextClick = { viewModel.processAction(OnboardingContract.Action.NextStep) },
         onBackClick = { viewModel.processAction(OnboardingContract.Action.PreviousStep) },
     )
 }
 
 @Composable
-fun OnboardingCompleteScreen2(
+fun OnboardingCompleteScreen1(
     state: OnboardingContract.State,
     onNextClick: () -> Unit,
     onBackClick: () -> Unit,
-    currentStep: Int = 0,
-    totalSteps: Int = 0,
 ) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(OrbitTheme.colors.gray_900)
-            .statusBarsPadding()
-            .navigationBarsPadding()
-            .imePadding(),
+    OnboardingScreen(
+        currentStep = 0,
+        totalSteps = 0,
+        isButtonEnabled = true,
+        onNextClick = onNextClick,
+        onBackClick = onBackClick,
+        showTopAppBar = true,
+        showTopAppBarActions = false,
+        buttonLabel = "다음",
     ) {
-        OnBoardingTopAppBar(
-            currentStep = currentStep,
-            totalSteps = totalSteps,
-            onBackClick = onBackClick,
-            showTopAppBarActions = false,
-        )
-
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .weight(1f),
-        ) {
+        Column(modifier = Modifier.fillMaxSize()) {
             Spacer(modifier = Modifier.heightForScreenPercentage(0.05f))
             Text(
                 text = stringResource(id = R.string.onboarding_completed_step1_subtitle),
@@ -87,30 +65,20 @@ fun OnboardingCompleteScreen2(
             )
             Spacer(modifier = Modifier.height(12.dp))
             Text(
-                text = stringResource(id = R.string.onboarding_completed_step2_title),
+                text = stringResource(id = R.string.onboarding_completed_step1_title),
                 style = OrbitTheme.typography.heading1SemiBold,
                 color = OrbitTheme.colors.white,
                 modifier = Modifier.fillMaxWidth(),
                 textAlign = TextAlign.Center,
             )
-
             Box(
                 modifier = Modifier.fillMaxSize(),
             ) {
                 LottieAnimation(
                     modifier = Modifier
-                        .scale(1.2f)
-                        .offset(y = (-70).dp),
-                    resId = core.designsystem.R.raw.step3,
-                )
-                OrbitButton(
-                    label = "시작하기",
-                    modifier = Modifier
-                        .padding(horizontal = 20.dp)
-                        .padding(bottom = 24.dp)
-                        .align(Alignment.BottomCenter),
-                    onClick = onNextClick,
-                    enabled = true,
+                        .scale(1.4f)
+                        .offset(y = (-50).dp),
+                    resId = core.designsystem.R.raw.step2,
                 )
             }
         }
@@ -119,9 +87,9 @@ fun OnboardingCompleteScreen2(
 
 @Composable
 @Preview
-fun OnboardingCompleteScreen2Preview() {
+fun OnboardingCompleteScreen1Preview() {
     OrbitTheme {
-        OnboardingCompleteScreen2(
+        OnboardingCompleteScreen1(
             state = OnboardingContract.State(),
             onNextClick = {},
             onBackClick = {},
