+++ feature/onboarding/src/main/java/com/yapp/onboarding/OnboardingCompleteScreen2.kt
@@ -1,16 +1,22 @@
-package com.kms.onboarding
+package com.yapp.onboarding
 
 import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.res.stringResource
@@ -19,42 +25,58 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.yapp.designsystem.theme.OrbitTheme
+import com.yapp.onboarding.component.OnBoardingTopAppBar
+import com.yapp.ui.component.button.OrbitButton
 import com.yapp.ui.component.lottie.LottieAnimation
 import com.yapp.ui.utils.heightForScreenPercentage
 import feature.onboarding.R
 
 @Composable
-fun OnboardingCompleteRoute(
+fun OnboardingCompleteRoute2(
     viewModel: OnboardingViewModel,
 ) {
     val state by viewModel.container.stateFlow.collectAsStateWithLifecycle()
     BackHandler {
         viewModel.processAction(OnboardingContract.Action.PreviousStep) // ✅ ViewModel에서 처리
     }
-    OnboardingCompleteScreen1(
+    OnboardingCompleteScreen2(
         state = state,
-        onNextClick = { viewModel.processAction(OnboardingContract.Action.NextStep) },
+        onNextClick = {
+            viewModel.processAction(OnboardingContract.Action.CompleteOnboarding)
+            viewModel.processAction(OnboardingContract.Action.CreateAlarm)
+        },
         onBackClick = { viewModel.processAction(OnboardingContract.Action.PreviousStep) },
     )
 }
 
 @Composable
-fun OnboardingCompleteScreen1(
+fun OnboardingCompleteScreen2(
     state: OnboardingContract.State,
     onNextClick: () -> Unit,
     onBackClick: () -> Unit,
+    currentStep: Int = 0,
+    totalSteps: Int = 0,
 ) {
-    OnboardingScreen(
-        currentStep = 0,
-        totalSteps = 0,
-        isButtonEnabled = true,
-        onNextClick = onNextClick,
-        onBackClick = onBackClick,
-        showTopAppBar = true,
-        showTopAppBarActions = false,
-        buttonLabel = "다음",
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(OrbitTheme.colors.gray_900)
+            .statusBarsPadding()
+            .navigationBarsPadding()
+            .imePadding(),
     ) {
-        Column(modifier = Modifier.fillMaxSize()) {
+        OnBoardingTopAppBar(
+            currentStep = currentStep,
+            totalSteps = totalSteps,
+            onBackClick = onBackClick,
+            showTopAppBarActions = false,
+        )
+
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .weight(1f),
+        ) {
             Spacer(modifier = Modifier.heightForScreenPercentage(0.05f))
             Text(
                 text = stringResource(id = R.string.onboarding_completed_step1_subtitle),
@@ -65,20 +87,30 @@ fun OnboardingCompleteScreen1(
             )
             Spacer(modifier = Modifier.height(12.dp))
             Text(
-                text = stringResource(id = R.string.onboarding_completed_step1_title),
+                text = stringResource(id = R.string.onboarding_completed_step2_title),
                 style = OrbitTheme.typography.heading1SemiBold,
                 color = OrbitTheme.colors.white,
                 modifier = Modifier.fillMaxWidth(),
                 textAlign = TextAlign.Center,
             )
+
             Box(
                 modifier = Modifier.fillMaxSize(),
             ) {
                 LottieAnimation(
                     modifier = Modifier
-                        .scale(1.4f)
-                        .offset(y = (-50).dp),
-                    resId = core.designsystem.R.raw.step2,
+                        .scale(1.2f)
+                        .offset(y = (-70).dp),
+                    resId = core.designsystem.R.raw.step3,
+                )
+                OrbitButton(
+                    label = "시작하기",
+                    modifier = Modifier
+                        .padding(horizontal = 20.dp)
+                        .padding(bottom = 24.dp)
+                        .align(Alignment.BottomCenter),
+                    onClick = onNextClick,
+                    enabled = true,
                 )
             }
         }
@@ -87,9 +119,9 @@ fun OnboardingCompleteScreen1(
 
 @Composable
 @Preview
-fun OnboardingCompleteScreen1Preview() {
+fun OnboardingCompleteScreen2Preview() {
     OrbitTheme {
-        OnboardingCompleteScreen1(
+        OnboardingCompleteScreen2(
             state = OnboardingContract.State(),
             onNextClick = {},
             onBackClick = {},
