+++ feature/onboarding/src/main/java/com/yapp/onboarding/OnboardingNavGraph.kt
@@ -1,4 +1,4 @@
-package com.kms.onboarding
+package com.yapp.onboarding
 
 import android.net.Uri
 import androidx.compose.runtime.LaunchedEffect
