+++ feature/onboarding/src/main/java/com/yapp/onboarding/OnboardingTimeOfBirthScreen.kt
@@ -1,4 +1,4 @@
-package com.kms.onboarding
+package com.yapp.onboarding
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
