+++ feature/onboarding/src/main/java/com/yapp/onboarding/OnboardingViewModel.kt
@@ -1,4 +1,4 @@
-package com.kms.onboarding
+package com.yapp.onboarding
 
 import android.util.Log
 import androidx.lifecycle.SavedStateHandle
